requirements.txt
# Core LLM library
openai>=1.7.0

# Vector Database (requires Pydantic v1)
chromadb>=0.4.22

# Local Embeddings (stable Windows pins)
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
sentence-transformers==2.6.1
transformers==4.39.3
accelerate==0.28.0

# LangChain (aligned versions - IMPORTANT)
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-openai==0.1.7
langsmith==0.1.31

# LlamaIndex (Agent Framework)
llama-index>=0.10.0
llama-index-agent-openai>=0.2.0
llama-index-llms-ollama>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-core>=0.10.0

# Gmail Integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# UI
streamlit>=1.30.0
fastapi>=0.110
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tqdm>=4.66.0

# Pin Pydantic v1 (IMPORTANT for Chroma 0.4.x)
pydantic>=1.10.17,<2.0.0
